feat: added problem/solution/result section to project detail page, along with updates on the image styling for the page
