+++ src/components/layout/projectImgs.js
@@ -1,5 +1,5 @@
 import React from 'react';
-// import Tooltip from '@material-ui/core/Tooltip';
+import Grid from '@material-ui/core/Grid';
 import './styles/projectImgs.scss';
 
 class ProjectImgs extends React.Component {
@@ -7,23 +7,19 @@ class ProjectImgs extends React.Component {
     const { modalData } = this.props;
     let detailPageInfo = this.props.data;
     return (
-      <div className="detail-img-container">
+      <Grid container spacing={4} className="detail-img-container">
         { detailPageInfo.images &&
-          (
-            <div>
-              {
-                detailPageInfo.images.map((image, key) => {
+          (<>
+              {detailPageInfo.images.map((image, key) => {
                   return (
-                    <div key={key} className="tooltip" onClick={() => modalData(image)}>
+                    <Grid item xs={12} sm={12} md={6} key={key} className="tooltip" onClick={() => modalData(image)}>
                       <img src={image.src} alt={image.alt} className="project-img" title="Click to enlarge photo"/>
-                    </div>
+                    </Grid>
                   );
-                })
-              }
-            </div>
-          )
+               })}
+            </>)
         }
-      </div>
+      </Grid>
     );
   }
 };
