+++ src/components/layout/styles/projectDetails.scss
@@ -58,6 +58,10 @@
   text-align: left;
 }
 
+.project-writeup {
+  padding: 2em 8vw;
+}
+
 .back-icon {
   vertical-align: middle;
 }
