+++ src/components/layout/styles/projectImgs.scss
@@ -7,10 +7,6 @@
   cursor: pointer;
 }
 
-.detail-img-container {
-  padding-right: 50px;
-}
-
 .modal-body:focus {
   outline: unset;
 }
